+++ crates/nu-explore/src/views/cursor/mod.rs
@@ -1,9 +1,9 @@
-mod windowcursor;
-mod xycursor;
+mod window_cursor;
+mod window_cursor_2d;
 
 use anyhow::{bail, Result};
-pub use windowcursor::WindowCursor;
-pub use xycursor::WindowCursor2D;
+pub use window_cursor::WindowCursor;
+pub use window_cursor_2d::WindowCursor2D;
 
 /// A 1-dimensional cursor to track a position from 0 to N
 ///
