+++ crates/nu-explore/src/views/cursor/window_cursor.rs
