+++ crates/nu-explore/src/views/cursor/window_cursor_2d.rs
